+++ occurrence-download/run-workflow-schema-migration.sh
@@ -24,6 +24,16 @@ echo "Assembling jar for $ENV"
 #Oozie uses timezone UTC
 mvn --settings profiles.xml -U -P$P -DskipTests -Duser.timezone=UTC clean install package assembly:single
 
+#Is any download running?
+IFS=$'\n'
+WFS=($(oozie jobs -oozie $OOZIE  -jobtype wf -filter "status=RUNNING;status=PREP;status=SUSPENDED;name=${ENV}-occurrence-download;name=${ENV}-species-list-download;name=${ENV}-create-tables"))
+unset IFS
+if [ ${#WFS[@]} > 1  ]; then
+  echo -e "$(tput setaf 1)Download workflow can not be installed while download or create HDFS table workflows are running!!$(tput sgr0) \n"
+  ( IFS=$'\n'; echo "${WFS[*]}" )
+  exit 1
+fi
+
 java -classpath "target/occurrence-download-workflows-$ENV/lib/*" org.gbif.occurrence.download.conf.DownloadConfBuilder $P  target/occurrence-download-workflows-$ENV/lib/occurrence-download.properties profiles.xml
 echo "Copy to hadoop"
 sudo -u hdfs hdfs dfs -rm -r -f /occurrence-download-workflows-new-schema-$ENV/
